Move source files to main module
