--- src/dev/weary/zomboid/plugin/ClassNodeTransformer.java
@@ -1,10 +1,0 @@
-package dev.weary.zomboid.plugin;
-
-
-import org.objectweb.asm.tree.ClassNode;
-
-@FunctionalInterface
-public interface ClassNodeTransformer {
-
-    void transformClass(ClassNode classNode);
-}
