--- src/dev/weary/zomboid/util/asm/AsmException.java
@@ -1,8 +1,0 @@
-package dev.weary.zomboid.util.asm;
-
-public class AsmException extends RuntimeException {
-
-    public AsmException(String message) {
-        super(message);
-    }
-}
