+++ src/main/java/dev/weary/zomboid/Debug.java
@@ -1,0 +1,11 @@
+package dev.weary.zomboid;
+
+import dev.weary.zomboid.agent.ui.UI;
+import zombie.Lua.Event;
+
+public class Debug {
+
+    public static void main(String[] args) {
+        UI.showWindow();
+    }
+}
