+++ src/main/java/dev/weary/zomboid/plugin/Transformer.java
@@ -1,0 +1,15 @@
+package dev.weary.zomboid.plugin;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Transformer {
+    boolean dumpClass() default false;
+    String dumpMethod() default "";
+    boolean generateDiff() default false;
+}
+
