+++ src/main/java/dev/weary/zomboid/util/asm/AsmException.java
@@ -1,0 +1,8 @@
+package dev.weary.zomboid.util.asm;
+
+public class AsmException extends RuntimeException {
+
+    public AsmException(String message) {
+        super(message);
+    }
+}
